requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
langdetect
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
pydantic
langdetect
torch
transformers
sentencepiece
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
google-generativeai
google-genai
